新創公司的公司資訊新增投資按鍵 (#344)
